* Number of children in a BvhNode is not limited to 2 anymore * Bvh building algo now collapses every secound layer into its parent, creating 4-child nodes * Added BvhToFile object to create a textual representation of the bvh for analysis

+++ Code/SPath/src/kaloffl/spath/scene/Camera.scala
@@ -12,8 +12,8 @@ class Camera(
     val position: Vec3d,
     val forward: Vec3d,
     val up: Vec3d,
-    val aperture: Float,
-    val focalLength: Float) {
+    val aperture: Double,
+    val focalLength: Double) {
 
   val right = forward.cross(up).normalize
